Create README.md
Greetings fellow developers and enthusiasts! 🌟 I'm a fresh recruit at 42 Madrid, and my journey has been a fusion of artistic exploration and technological curiosity. I've spent a significant part of my life immersed in the realms of digital photography, painting, and the intricate tapestry of multimedia content creation, even venturing into the world of music composition.
Navigating the Web Through Time

My initial exposure to the world of programming dates back to the late '90s. In those days, I was already crafting web pages using the foundational HTML, evolving later into a specialization in Flash. Despite the unfortunate decline of Flash technology, which I was diligently studying, I redirected my creative energies toward refining my skills in photography and painting.
Rediscovering the Digital Landscape

A later chapter in my digital journey led me to rediscover the web through platforms like WordPress. However, a new dawn has arrived, and I'm eagerly embracing the opportunity to learn and evolve. Today, my focus is set on becoming a seasoned professional in the dynamic field of virtual and augmented reality application development.
Unveiling Dreams in Virtual Reality

Throughout my life, I've harbored a dream – a dream where the immersive experience of virtual and augmented reality becomes an integral part of our everyday lives. Today, armed with a passion for creation, I'm on a mission to transform this dream into reality by acquiring the skills needed to craft captivating and immersive digital content.
GitHub: A Canvas for Collaboration

On GitHub, I present to you my ongoing journey. This is not just a collection of code; it's a canvas where artistry meets technology. I invite you to explore my projects, share insights, and join me on this exciting path of continuous learning and innovation.
Future Collaborations

I'm not just here to share my work; I'm here to collaborate and contribute. I believe in the power of collective creativity, and I look forward to connecting with like-minded individuals who share a passion for pushing the boundaries of what's possible in the realms of virtual and augmented reality.

Let's build the future together, one line of code at a time! 🚀